+++ src/entrypoint.c
@@ -11,12 +11,14 @@ SEXP wrap__cal_sheet_names(SEXP path);
 SEXP wrap__cal_read_sheet(SEXP path, SEXP sheet);
 SEXP wrap__cal_read_sheet_df(SEXP path, SEXP sheet, SEXP col_names, SEXP skip, SEXP fill_merged);
 SEXP wrap__cal_sheet_dims(SEXP path, SEXP sheet);
+SEXP wrap__cal_merge_regions(SEXP path, SEXP sheet);
 
 static const R_CallMethodDef CallEntries[] = {
     {"wrap__cal_sheet_names", (DL_FUNC) &wrap__cal_sheet_names, 1},
     {"wrap__cal_read_sheet", (DL_FUNC) &wrap__cal_read_sheet, 2},
     {"wrap__cal_read_sheet_df", (DL_FUNC) &wrap__cal_read_sheet_df, 5},
     {"wrap__cal_sheet_dims", (DL_FUNC) &wrap__cal_sheet_dims, 2},
+    {"wrap__cal_merge_regions", (DL_FUNC) &wrap__cal_merge_regions, 2},
     {NULL, NULL, 0}
 };